debug

+++ compass.c
@@ -93,7 +93,7 @@ int compass_read_scatter(Cmps *data)
 	   {
 	        printf("Compass write register:mode_reg\n");
 	   }*/
-	int val = wiringPiI2CReadReg8(fd, 0x09);
+	uint8_t val = wiringPiI2CReadReg8(fd, 0x09);
 	printf("status = %d\n", val);
 	data->x_value = read_out(fd, x_msb_reg, x_lsb_reg);
 	data->y_value = read_out(fd, y_msb_reg, y_lsb_reg);
